+++ frontend/src/App.jsx
@@ -7,7 +7,7 @@ import Footer from './components/Footer';
 import History from './pages/History';
 import Overview from './pages/Overview';
 import Analysis from './pages/Analysis';
-import Build from './pages/Build';
+import BuildHistory from './pages/BuildHistory';
 import Frontpage from './pages/Frontpage';
 import Team from './pages/Team';
 import Suite from './pages/Suite';
@@ -53,7 +53,7 @@ const App = () => {
                                 <Suite />
                             </Route>
                             <Route path="/series/:seriesId/build/:buildId/history">
-                                <Build />
+                                <BuildHistory />
                             </Route>
                             <Route path="/series/:seriesId/history">
                                 <History />
